separated the logic for preparing podList array into a custom hook, from PodGrid component
