+++ src/components/HeatMapComponents/PodGrid.jsx
@@ -20,6 +20,7 @@ import QueryTimeWindowConfiguration from "./QueryTimeWindowConfiguration";
 import PodSelector from "./PodSelector";
 import PodSorter from "./PodSorter";
 import PodFilter from "./PodFilter";
+import usePodListProcessor from "../../hooks/usePodListProcessor";
 
 const PodGrid = ({
   defaultView,
@@ -42,68 +43,22 @@ const PodGrid = ({
   const [filterConfig, setFilterConfig] = useState({
     type: "",
     value: "",
-    operator: "over",
+  });
+
+  // Funan: separated the preparation of the pod list into a custom hook, usePodListProcessor
+  const processedPodList = usePodListProcessor({
+    podStatuses,
+    cpuUsageOneValue,
+    memoryUsageOneValue,
+    latencyAppRequestOneValue,
+    selectedMetric,
+    filterConfig,
+    metricToSort,
+    defaultView,
   });
 
   if (!podStatuses.allPodsStatus) {
     return <div>loading...</div>;
-  }
-
-  const podList = podStatuses?.allPodsStatus?.map((pod) => {
-    const podObj = {
-      podName: pod.podName,
-      namespace: pod.namespace,
-      status: pod.status,
-      readiness: pod.readiness,
-      containers: pod.containers,
-      service: pod.service,
-      deploymentName: pod.deploymentName,
-      selectedMetric,
-    };
-
-    const cpuData = cpuUsageOneValue?.resourceUsageOneValue?.find(
-      (obj) => obj.name === pod.podName,
-    );
-    podObj.cpuDataRelative = cpuData?.usageRelativeToRequest;
-    podObj.cpuDataAbsolute = cpuData?.usageAbsolute;
-
-    const memoryData = memoryUsageOneValue?.resourceUsageOneValue?.find(
-      (obj) => obj.name === pod.podName,
-    );
-    podObj.memoryDataRelative = memoryData?.usageRelativeToRequest;
-    podObj.memoryDataAbsolute = memoryData?.usageAbsolute;
-
-    const latencyData =
-      latencyAppRequestOneValue?.latencyAppRequestOneValue?.find(
-        (obj) => obj.name === pod.podName,
-      );
-    podObj.latencyData = latencyData?.avgCombinedLatency;
-
-    return podObj;
-  });
-
-  let processedPodList = [...podList];
-
-  // Apply filter
-  if (filterConfig.type && filterConfig.value && !defaultView) {
-    processedPodList = processedPodList.filter((pod) => {
-      if (["cpuRelative", "memoryRelative"].includes(filterConfig.type)) {
-        const value = pod[filterConfig.type];
-        const threshold = parseFloat(filterConfig.value);
-        return filterConfig.operator === "over"
-          ? value > threshold
-          : value < threshold;
-      }
-      return pod[filterConfig.type] === filterConfig.value;
-    });
-  }
-
-  // Only sort if a metricToSort is selected and default view is false
-  if (metricToSort && !defaultView) {
-    processedPodList.sort(
-      // when metric data is not available, default to 0
-      (a, b) => (b[metricToSort] || 0) - (a[metricToSort] || 0),
-    );
   }
 
   const buttonArray = processedPodList.map((podObj) => (
@@ -116,7 +71,7 @@ const PodGrid = ({
         clickedPod.podName === podObj.podName &&
         clickedPod.namespace === podObj.namespace &&
         clickedPod.containers === podObj.containers
-        // && clickedPod.deploymentName === podObj.deploymentName // don't need to check deployment for isClicked, will mess up with selector
+        // && clickedPod.deploymentName === podObj.deploymentName // Don't need to check deployment, it also messes up the selector
       }
       onClick={() => {
         setClickedPod({
@@ -153,7 +108,7 @@ const PodGrid = ({
           backendUrl={backendUrl}
         />
         <PodSelector
-          podList={podList}
+          podList={processedPodList}
           setClickedPod={setClickedPod}
           defaultView={defaultView}
           setDefaultView={setDefaultView}
@@ -165,7 +120,7 @@ const PodGrid = ({
           setDefaultView={setDefaultView}
         />
         <PodFilter
-          podList={podList}
+          podList={processedPodList}
           setFilterConfig={setFilterConfig}
           defaultView={defaultView}
           setDefaultView={setDefaultView}
@@ -198,13 +153,10 @@ const PodGrid = ({
               });
               setSelectedMetric("cpu");
               setQueryTimeWindow("1m");
-              // Reset filter configuration
               setFilterConfig({
                 type: "",
                 value: "",
-                operator: "over",
               });
-              // Reset sort configuration
               setMetricToSort("");
             }}
             className="rounded-2xl border-4 border-blue-600 bg-gradient-to-r from-slate-200 to-slate-100 px-2 py-4 text-lg font-semibold text-blue-600 transition duration-200 hover:border-2 hover:bg-gradient-to-r hover:from-[#1d4ed8] hover:to-[#2563eb] hover:text-slate-100"
